feat: implement routing and layout structure with Navbar and Footer components
