+++ package.json
@@ -24,6 +24,8 @@
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-hook-form": "^7.66.0",
+    "react-router": "^7.9.5",
+    "react-router-dom": "^7.9.5",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.16",
     "zod": "^4.1.12"
