+++ src/pages/about.pages.tsx
@@ -1,0 +1,8 @@
+
+const aboutPages = () => {
+  return (
+    <div>aboutPages</div>
+  )
+}
+
+export default aboutPages
