+++ src/pages/landing.pages.tsx
@@ -1,0 +1,8 @@
+
+const landingPages = () => {
+  return (
+    <div>landingPages</div>
+  )
+}
+
+export default landingPages
