+++ src/pages/services.pages.tsx
@@ -1,0 +1,8 @@
+
+const servicesPages = () => {
+  return (
+    <div>servicesPages</div>
+  )
+}
+
+export default servicesPages
